backend/requirements.txt
Flask
psycopg2-binary
PyJWT
flask-cors
